remove support for has_and_belongs_to_many associations
Using `has_and_belongs_to_many` would now raise an exception.

Support for this assosciation has to be removed because of the errors
raised from it as mentioned in #1. Couldn't find how and why the issue
is caused in particular. Probably, because of some changes in
ActiveRecord. So, decided to remove support for it.